+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Handle storage retrieval failures by falling back to default dimmer settings.
+- Fix schedule retrieval so local data is used when sync storage runs out of quota.
 
 ## [1.4.0] - 2025-11-02